refactored auth and added comments

+++ app/controllers/notes_controller.rb
@@ -2,16 +2,20 @@ class NotesController < ApplicationController
     def create
         note = Note.new(note_params)
         if note.save
-            render json: {note: NoteSerializer.new(note)}
+            render json: note, status: :created
         else
-            render json: {errors: note.errors.full_messages}
+            render json: {errors: note.errors.full_messages}, status: :not_acceptable
         end
     end
 
     def destroy 
-        note = Note.find(params[:id])
-        note.destroy
-        render json: note
+        note = Note.find_by(id: params[:id])
+        if note
+            note.destroy
+            render json: note
+        else
+            render json: {errors: 'No note found'}, status: :not_found
+        end
     end
 
     private
